requirements.txt
streamlit
openai
langchain
apify-client
scikit-learn
python-dotenv
